requirements.txt
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

# Google Cloud Storage
google-cloud-storage>=2.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
click>=8.1.0
pyyaml>=6.0
pytz>=2023.3

# Image processing
opencv-python>=4.8.0

# Network analysis (if needed)
networkx>=3.1

# Protobuf (for serialization)
protobuf>=4.23.0

# Template engine
jinja2>=3.1.0

# Parsing utilities
pyparsing>=3.1.0

# Math/symbolic
sympy>=1.12

# Web framework (if needed)
werkzeug>=2.3.0

# Build tools
wheel>=0.40.0
setuptools>=68.0.0

# Code quality
pyflakes>=3.0.0

# Additional utilities
six>=1.16.0
tqdm>=4.65.0  # Progress bars for training